requirements.txt
openai>=1.95.1
pydantic>=2.11.7
asyncio>=3.4.3
numpy>=1.24.0
typing-extensions>=4.14.1
pyyaml>=6.0
python-dotenv>=1.0.0